Use cache for the throttle definitions
